+++ examples/1d/elixir_eulermulti_es.jl
@@ -3,20 +3,22 @@ using OrdinaryDiffEq
 using Trixi
 
 ###############################################################################
-# semidiscretization of the Lattice-Boltzmann equations for the D2Q9 scheme
+# semidiscretization of the compressible Euler multicomponent equations
+equations = CompressibleEulerMulticomponentEquations1D(gammas        = (1.4, 1.4),
+                                                       gas_constants = (0.4, 0.4))
 
-equations = LatticeBoltzmannEquations2D(Ma=0.1, Re=Inf)
 
-initial_condition = initial_condition_constant
+initial_condition = initial_condition_weak_blast_wave
 
-surface_flux = flux_godunov
-solver = DGSEM(3, surface_flux)
+surface_flux = flux_lax_friedrichs
+volume_flux  = flux_chandrashekar
+solver = DGSEM(3, surface_flux, VolumeIntegralFluxDifferencing(volume_flux))
 
-coordinates_min = (-1, -1)
-coordinates_max = ( 1,  1)
+coordinates_min = (-2,)
+coordinates_max = ( 2,)
 mesh = TreeMesh(coordinates_min, coordinates_max,
-                initial_refinement_level=4,
-                n_cells_max=10_000,)
+                initial_refinement_level=5,
+                n_cells_max=10_000)
 
 
 semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
@@ -25,13 +27,15 @@ semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
 ###############################################################################
 # ODE solvers, callbacks etc.
 
-tspan = (0.0, 1.0)
+tspan = (0.0, 0.4)
 ode = semidiscretize(semi, tspan)
 
 summary_callback = SummaryCallback()
 
 analysis_interval = 100
-analysis_callback = AnalysisCallback(semi, interval=analysis_interval)
+
+analysis_callback = AnalysisCallback(semi, interval=analysis_interval,
+                                     extra_analysis_integrals=(Trixi.density,))
 
 alive_callback = AliveCallback(analysis_interval=analysis_interval)
 
@@ -41,17 +45,14 @@ save_restart = SaveRestartCallback(interval=100,
 save_solution = SaveSolutionCallback(interval=100,
                                      save_initial_solution=true,
                                      save_final_solution=true,
-                                     solution_variables=cons2macroscopic)
+                                     solution_variables=:primitive)
 
-stepsize_callback = StepsizeCallback(cfl=1.0)
-
-collision_callback = LBMCollisionCallback()
+stepsize_callback = StepsizeCallback(cfl=0.8)
 
 callbacks = CallbackSet(summary_callback,
-                        analysis_callback, alive_callback,
+                        analysis_callback, alive_callback, 
                         save_restart, save_solution,
-                        stepsize_callback,
-                        collision_callback)
+                        stepsize_callback)
 
 
 ###############################################################################