+++ examples/1d/elixir_mhdmulti_eoc.jl
@@ -3,20 +3,22 @@ using OrdinaryDiffEq
 using Trixi
 
 ###############################################################################
-# semidiscretization of the Lattice-Boltzmann equations for the D3Q27 scheme
+# semidiscretization of the compressible ideal GLM-MHD equations
 
-equations = LatticeBoltzmannEquations3D(Ma=0.1, Re=Inf)
+equations = IdealGlmMhdMulticomponentEquations1D(gammas         = (5/3, 5/3, 5/3),
+                                                 gas_constants  = (2.08, 2.08, 2.08))
 
-initial_condition = initial_condition_constant
+initial_condition = initial_condition_convergence_test
 
-surface_flux = flux_godunov
-solver = DGSEM(3, surface_flux)
+surface_flux = flux_lax_friedrichs
+volume_flux  = flux_derigs_etal
+solver = DGSEM(3, surface_flux, VolumeIntegralFluxDifferencing(volume_flux))
 
-coordinates_min = (-1, -1, -1)
-coordinates_max = ( 1,  1,  1)
+coordinates_min = 0.0
+coordinates_max = 1.0
 mesh = TreeMesh(coordinates_min, coordinates_max,
-                initial_refinement_level=3,
-                n_cells_max=10_000,)
+                initial_refinement_level=2,
+                n_cells_max=10_000)
 
 
 semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
@@ -25,33 +27,29 @@ semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
 ###############################################################################
 # ODE solvers, callbacks etc.
 
-tspan = (0.0, 1.0)
+tspan = (0.0, 2.0)
 ode = semidiscretize(semi, tspan)
 
 summary_callback = SummaryCallback()
 
 analysis_interval = 100
-analysis_callback = AnalysisCallback(semi, interval=analysis_interval)
+analysis_callback = AnalysisCallback(semi, interval=analysis_interval, save_analysis=true)
 
 alive_callback = AliveCallback(analysis_interval=analysis_interval)
 
-save_restart = SaveRestartCallback(interval=100,
-                                   save_final_restart=true)
-
-save_solution = SaveSolutionCallback(interval=100,
+save_solution = SaveSolutionCallback(interval=10,
                                      save_initial_solution=true,
                                      save_final_solution=true,
-                                     solution_variables=cons2macroscopic)
+                                     solution_variables=cons2prim)
 
-stepsize_callback = StepsizeCallback(cfl=1.0)
-
-collision_callback = LBMCollisionCallback()
+cfl = 0.1
+stepsize_callback = StepsizeCallback(cfl=cfl)
 
 callbacks = CallbackSet(summary_callback,
-                        analysis_callback, alive_callback,
-                        save_restart, save_solution,
-                        stepsize_callback,
-                        collision_callback)
+                        analysis_callback,
+                        alive_callback,
+                        save_solution,
+                        stepsize_callback)
 
 
 ###############################################################################