+++ examples/1d/elixir_mhdmulti_es.jl
@@ -3,20 +3,21 @@ using OrdinaryDiffEq
 using Trixi
 
 ###############################################################################
-# semidiscretization of the Lattice-Boltzmann equations for the D2Q9 scheme
+# semidiscretization of the ideal MHD equations
+equations = IdealGlmMhdMulticomponentEquations1D(gammas           = (2.0, 2.0, 2.0),
+                                                 gas_constants    = (2.0, 2.0, 2.0))
 
-equations = LatticeBoltzmannEquations2D(Ma=0.1, Re=Inf)
+initial_condition = initial_condition_briowu_shock_tube
 
-initial_condition = initial_condition_constant
+surface_flux = flux_lax_friedrichs
+volume_flux  = flux_derigs_etal
+solver = DGSEM(3, surface_flux, VolumeIntegralFluxDifferencing(volume_flux))
 
-surface_flux = flux_godunov
-solver = DGSEM(3, surface_flux)
-
-coordinates_min = (-1, -1)
-coordinates_max = ( 1,  1)
+coordinates_min = 0
+coordinates_max = 1
 mesh = TreeMesh(coordinates_min, coordinates_max,
                 initial_refinement_level=4,
-                n_cells_max=10_000,)
+                n_cells_max=10_000)
 
 
 semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
@@ -25,13 +26,15 @@ semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
 ###############################################################################
 # ODE solvers, callbacks etc.
 
-tspan = (0.0, 1.0)
+tspan = (0.0, 0.04)
 ode = semidiscretize(semi, tspan)
 
 summary_callback = SummaryCallback()
 
 analysis_interval = 100
-analysis_callback = AnalysisCallback(semi, interval=analysis_interval)
+analysis_callback = AnalysisCallback(semi, interval=analysis_interval,
+                                     extra_analysis_errors=(:l2_error_primitive,
+                                                            :linf_error_primitive))
 
 alive_callback = AliveCallback(analysis_interval=analysis_interval)
 
@@ -41,17 +44,14 @@ save_restart = SaveRestartCallback(interval=100,
 save_solution = SaveSolutionCallback(interval=100,
                                      save_initial_solution=true,
                                      save_final_solution=true,
-                                     solution_variables=cons2macroscopic)
+                                     solution_variables=cons2prim)
 
-stepsize_callback = StepsizeCallback(cfl=1.0)
-
-collision_callback = LBMCollisionCallback()
+stepsize_callback = StepsizeCallback(cfl=0.5)
 
 callbacks = CallbackSet(summary_callback,
                         analysis_callback, alive_callback,
                         save_restart, save_solution,
-                        stepsize_callback,
-                        collision_callback)
+                        stepsize_callback)
 
 
 ###############################################################################