+++ examples/2d/elixir_mhdmulti_ec.jl
@@ -3,20 +3,22 @@ using OrdinaryDiffEq
 using Trixi
 
 ###############################################################################
-# semidiscretization of the Lattice-Boltzmann equations for the D2Q9 scheme
+# semidiscretization of the compressible ideal GLM-MHD equations
 
-equations = LatticeBoltzmannEquations2D(Ma=0.1, Re=Inf)
+equations = IdealGlmMhdMulticomponentEquations2D(gammas         = (1.4, 1.4),
+                                                 gas_constants  = (1.0, 1.0))
 
-initial_condition = initial_condition_constant
+initial_condition = initial_condition_weak_blast_wave
 
-surface_flux = flux_godunov
-solver = DGSEM(3, surface_flux)
+surface_flux = flux_derigs_etal
+volume_flux  = flux_derigs_etal
+solver = DGSEM(3, surface_flux, VolumeIntegralFluxDifferencing(volume_flux))
 
-coordinates_min = (-1, -1)
-coordinates_max = ( 1,  1)
+coordinates_min = (-2, -2)
+coordinates_max = ( 2,  2)
 mesh = TreeMesh(coordinates_min, coordinates_max,
                 initial_refinement_level=4,
-                n_cells_max=10_000,)
+                n_cells_max=10_000)
 
 
 semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
@@ -25,7 +27,7 @@ semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
 ###############################################################################
 # ODE solvers, callbacks etc.
 
-tspan = (0.0, 1.0)
+tspan = (0.0, 0.4)
 ode = semidiscretize(semi, tspan)
 
 summary_callback = SummaryCallback()
@@ -35,23 +37,22 @@ analysis_callback = AnalysisCallback(semi, interval=analysis_interval)
 
 alive_callback = AliveCallback(analysis_interval=analysis_interval)
 
-save_restart = SaveRestartCallback(interval=100,
-                                   save_final_restart=true)
-
-save_solution = SaveSolutionCallback(interval=100,
+save_solution = SaveSolutionCallback(interval=10,
                                      save_initial_solution=true,
                                      save_final_solution=true,
-                                     solution_variables=cons2macroscopic)
+                                     solution_variables=cons2prim)
 
-stepsize_callback = StepsizeCallback(cfl=1.0)
+cfl = 1.0
+stepsize_callback = StepsizeCallback(cfl=cfl)
 
-collision_callback = LBMCollisionCallback()
+glm_speed_callback = GlmSpeedCallback(glm_scale=0.5, cfl=cfl)
 
 callbacks = CallbackSet(summary_callback,
-                        analysis_callback, alive_callback,
-                        save_restart, save_solution,
+                        analysis_callback,
+                        alive_callback,
+                        save_solution,
                         stepsize_callback,
-                        collision_callback)
+                        glm_speed_callback)
 
 
 ###############################################################################